add editor assembly definition, move all editor scripts to one location (#882)
