+++ Assets/Editor/Atmospherics.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 71660cef1c540784e9b8cfb18c6966af
+guid: bfe77e10555311045abbf97b014c04c4
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
