+++ Assets/Editor/EditorDefinitions.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: adfd753e6e0bb234fb7143be344f1fa4
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
