+++ Assets/Editor/Inventory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4fbeea1b963ac50498bfc20d4ca45cdb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
