+++ Assets/Editor/Inventory/ContainerDescriptorEditor.cs
