--- Assets/Engine/Atmospherics/Editor.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-<<<<<<< Updated upstream:Assets/Content/Furniture/Machines/Atmospherics/Canister.meta
-guid: 12438234a7864ee4e9a7c4fae4949455
-=======
-guid: 04c771012d81cd641a69e02f6ace072e
->>>>>>> Stashed changes:Assets/Engine/Atmospherics/Editor.meta
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
