+++ Assets/Engine/Debug/EasyButtons/Editor.meta
@@ -1,8 +1,3 @@
-fileFormatVersion: 2
-guid: a1d0732956bd8c545bfa3bc82289f4f9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: fc7cddccfc8d4df981e373c980a34829
+timeCreated: 1641283253
