+++ Assets/Engine/Debug/EasyButtons/Editor/EasyButtons.asmref
@@ -1,0 +1,3 @@
+{
+    "reference": "GUID:adfd753e6e0bb234fb7143be344f1fa4"
+}
