+++ Assets/Engine/Debug/EasyButtons/Editor/EasyButtons.asmref.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: e1023e4e342a4d247a43239f791d5e7c
-folderAsset: yes
-DefaultImporter:
+guid: 9f0771fe28cfd4a48aab911f4dba7d20
+AssemblyDefinitionReferenceImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
